extract protobuf serialization from OperationMessage to Operation
