+++ syft/execution/operation.py
@@ -1,10 +1,10 @@
 import syft as sy
 from syft.workers.abstract import AbstractWorker
 
-# from syft.frameworks.torch.tensors.interpreters.placeholder import PlaceHolder
+from syft.frameworks.torch.tensors.interpreters.placeholder import PlaceHolder
 
-# from syft_proto.execution.v1.operation_pb2 import Operation as OperationPB
-# from syft_proto.types.syft.v1.arg_pb2 import Arg as ArgPB
+from syft_proto.execution.v1.operation_pb2 import Operation as OperationPB
+from syft_proto.types.syft.v1.arg_pb2 import Arg as ArgPB
 
 
 class Operation:
@@ -102,3 +102,151 @@ class Operation:
         cmd_kwargs = detailed_msg[3]
 
         return Operation(cmd_name, cmd_owner, cmd_args, cmd_kwargs, detailed_ids)
+
+
+    @staticmethod
+    def bufferize(worker: AbstractWorker, operation: "Operation") -> "OperationPB":
+        """
+        This function takes the attributes of a Operation and saves them in Protobuf
+        Args:
+            worker (AbstractWorker): a reference to the worker doing the serialization
+            operation (Operation): an Operation
+        Returns:
+            protobuf_obj: a Protobuf message holding the unique attributes of the message
+        Examples:
+            data = bufferize(message)
+        """
+        protobuf_op = OperationPB()
+        protobuf_op.command = operation.cmd_name
+
+        if type(operation.cmd_owner) == sy.generic.pointers.pointer_tensor.PointerTensor:
+            protobuf_owner = protobuf_op.owner_pointer
+        elif (
+            type(operation.cmd_owner)
+            == sy.frameworks.torch.tensors.interpreters.placeholder.PlaceHolder
+        ):
+            protobuf_owner = protobuf_op.owner_placeholder
+        else:
+            protobuf_owner = protobuf_op.owner_tensor
+
+        if operation.cmd_owner is not None:
+            protobuf_owner.CopyFrom(sy.serde.protobuf.serde._bufferize(worker, operation.cmd_owner))
+
+        if operation.cmd_args:
+            protobuf_op.args.extend(Operation._bufferize_args(worker, operation.cmd_args))
+
+        if operation.cmd_kwargs:
+            for key, value in operation.cmd_kwargs.items():
+                protobuf_op.kwargs.get_or_create(key).CopyFrom(
+                    Operation._bufferize_arg(worker, value)
+                )
+
+        if operation.return_ids is not None:
+            if type(operation.return_ids) == PlaceHolder:
+                return_ids = list((operation.return_ids,))
+            else:
+                return_ids = operation.return_ids
+
+            for return_id in return_ids:
+                if type(return_id) == PlaceHolder:
+                    protobuf_op.return_placeholders.extend(
+                        [sy.serde.protobuf.serde._bufferize(worker, return_id)]
+                    )
+                else:
+                    sy.serde.protobuf.proto.set_protobuf_id(protobuf_op.return_ids.add(), return_id)
+
+        return protobuf_op
+
+
+    @staticmethod
+    def unbufferize(
+        worker: AbstractWorker, protobuf_obj: "OperationPB"
+    ) -> "Operation":
+        """
+        This function takes the Protobuf version of this message and converts
+        it into an Operation. The bufferize() method runs the inverse of this method.
+
+        Args:
+            worker (AbstractWorker): a reference to the worker necessary for detailing. Read
+                syft/serde/serde.py for more information on why this is necessary.
+            protobuf_obj (OperationPB): the Protobuf message
+
+        Returns:
+            obj (Operation): an Operation
+
+        Examples:
+            message = unbufferize(sy.local_worker, protobuf_msg)
+        """
+        command = protobuf_obj.command
+        protobuf_owner = protobuf_obj.WhichOneof("owner")
+        if protobuf_owner:
+            owner = sy.serde.protobuf.serde._unbufferize(
+                worker, getattr(protobuf_obj, protobuf_obj.WhichOneof("owner"))
+            )
+        else:
+            owner = None
+        args = Operation._unbufferize_args(worker, protobuf_obj.args)
+
+        kwargs = {}
+        for key in protobuf_obj.kwargs:
+            kwargs[key] = Operation._unbufferize_arg(
+                worker, protobuf_obj.kwargs[key]
+            )
+
+        return_ids = [
+            sy.serde.protobuf.proto.get_protobuf_id(pb_id)
+            for pb_id in protobuf_obj.return_ids
+        ]
+
+        return_placeholders = [
+            sy.serde.protobuf.serde._unbufferize(worker, placeholder)
+            for placeholder in protobuf_obj.return_placeholders
+        ]
+
+        if return_placeholders:
+            if len(return_placeholders) == 1:
+                operation = Operation(
+                    command, owner, tuple(args), kwargs, return_placeholders[0]
+                )
+            else:
+                operation = Operation(
+                    command, owner, tuple(args), kwargs, return_placeholders
+                )
+        else:
+            operation = Operation(command, owner, tuple(args), kwargs, tuple(return_ids))
+
+        return operation
+
+    @staticmethod
+    def _bufferize_args(worker: AbstractWorker, args: list) -> list:
+        protobuf_args = []
+        for arg in args:
+            protobuf_args.append(Operation._bufferize_arg(worker, arg))
+        return protobuf_args
+
+    @staticmethod
+    def _bufferize_arg(worker: AbstractWorker, arg: object) -> ArgPB:
+        protobuf_arg = ArgPB()
+        try:
+            setattr(protobuf_arg, "arg_" + type(arg).__name__.lower(), arg)
+        except:
+            getattr(protobuf_arg, "arg_" + type(arg).__name__.lower()).CopyFrom(
+                sy.serde.protobuf.serde._bufferize(worker, arg)
+            )
+        return protobuf_arg
+
+    @staticmethod
+    def _unbufferize_args(worker: AbstractWorker, protobuf_args: list) -> list:
+        args = []
+        for protobuf_arg in protobuf_args:
+            args.append(Operation._unbufferize_arg(worker, protobuf_arg))
+        return args
+
+    @staticmethod
+    def _unbufferize_arg(worker: AbstractWorker, protobuf_arg: ArgPB) -> object:
+        protobuf_arg_field = getattr(protobuf_arg, protobuf_arg.WhichOneof("arg"))
+        try:
+            arg = sy.serde.protobuf.serde._unbufferize(worker, protobuf_arg_field)
+        except:
+            arg = protobuf_arg_field
+        return arg
